pyproject.toml
[project]
name = "shakespeare-x-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.4",
    "langgraph>=0.5.2",
    "pymysql>=1.1.1",
    "snowflake-id>=1.0.2",
    "sse-starlette>=2.4.1",
    "uvicorn>=0.35.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
